feat(coco-rs): TS-parity overhaul + multi-provider prompt foundation
Squash of 12 commits since 48d3d49c (Feat/features #91), covering the
TS-parity overhaul merged as #92, per-model builtin base_instructions,
the provider prompt role architecture target design, the first impl
slice of that design, and a round of macOS test-isolation fixes.

— TS-parity overhaul (PR #92)
- core: /rewind picker reaches MessageSelector parity; vercel-ai seam
  collapsed to services/inference (Message family relocated to
  coco-messages); SessionRuntime unifies TUI/SDK runner wiring;
  engine.rs split 4511 → 1550 LoC across builder/session/compaction/
  finalize_turn/prompt; workspace clippy-zero + Stop-hook + pre-commit
  gate.
- memory: port auto-memory + auto-dream (extract/dream/session under
  MemoryRuntime, gated by Feature::AutoMemory + ModelRole::Memory);
  turn-end fan-out from QueryEngine; /dream + /summary sentinels;
  SideQueryAdapter for the recall ranker.
- agentteam: extract 21 swarm modules into coco-coordinator (root
  layer); spawn-time identity (definition.model_role + per-spawn
  precedence + ThinkingLevel + SubagentRuntimeSnapshot for cache
  parity); SpawnMode {Fresh,Fork,Resume}; background AgentTool drives
  the engine, persists per-agent JSONL, opt-in periodic AgentSummary;
  model-driven resume via SendMessageTool; TUI permission bridge for
  Ask; mailbox/tmux/handoff TS-parity; per-agent MEMORY.md / hooks /
  skills / MCP inheritance; /agents slash command;
  Feature::AgentTeams Experimental.
- compact: Round-10 normalize pipeline (5 ordered passes, smoosh-
  before-sanitize, request-id-keyed merge, image-strip in tool-result
  arrays); post-compact plan-mode + async-agent attachments;
  RecompactionInfo through CompactRunOptions; SessionStart hooks after
  LLM compact; Tool Result Budget Phase 0 + Level-1 + Level-2 spill.
- skills/commands/plugins: Round-11 — bundled inventory + per-process
  nonce with O_NOFOLLOW|O_EXCL; TS-faithful prompt token renderer;
  CommandResult {Compact,Prompt,OpenDialog,Skip} + DialogSpec;
  PluginId/PluginScope, DFS dependency resolver, NFKD homograph fold,
  ZIP/mcpb loader with sha256 cache + traversal guard, 24-variant
  PluginError.
- @-mentions: parser fragment + agent-suffix + symbol fixes;
  MentionType::McpResource; TUI matcher → coco-file-search (nucleo +
  git ls-files / ripgrep) with SharedFileIndex pre-warm +
  git_index_watcher; auto-quoting / dir-trailing-slash / agent-prefix
  in accept_suggestion; nested-CLAUDE.md mention resolution.
- conventions: Config & Feature Gates rules in root CLAUDE.md.

— Builtin per-model base_instructions
- New common/config/src/instructions/ with 4 markdown prompts
  (gpt5_4 / gpt5_5 / gpt5_3_codex / gemini) embedded via
  include_str!; seed_builtin_models wires them into
  PartialModelInfo.base_instructions for the 5 known model ids.
- build_model_registry resolves base_instructions_file (relative to
  ~/.coco/) at build time; file always wins over inline within the
  merged entry. ConfigError::BaseInstructionsRead on IO failure.
- CLI: build_system_prompt_for_model so headless / SDK / TUI all
  resolve identity through runtime_config.model_registry; falls back
  to default Coco identity when no model is registered.

— Provider prompt role architecture (target design)
- New docs/coco-rs/provider-prompt-role-architecture.md, owned by the
  coco-context + coco-query + coco-inference + vercel-ai-* crates.
  Sources of truth: codex-rs/codex-api (OpenAI Responses),
  claude-code-kim (Anthropic Messages), gemini-cli (Gemini
  GenerateContent). 1725 lines covering envelope shape, (kind,family)
  routing, layout-adapter ownership in services/inference, native
  slot transport via provider_options["prompt_layout"], cache hash
  partitioning, warning bubbling, function-call adjacency, and 35
  enforced invariants. Includes a §0 TL;DR with the decision matrix
  (rejected alternatives + one-line rationale per choice).
- Replaces no-extension fields on LanguageModelV4CallOptions: payload
  rides on the existing ProviderOptions namespace mechanism. Provider
  crates parse "prompt_layout" via local serde mirror structs and do
  NOT depend on coco-inference. extra_body_hash excludes
  "prompt_layout" so prompt content cannot double-count as request-
  shaping extras.

— First impl slice of the prompt-role architecture
- vercel-ai-provider: System carries Vec<UserContentPart>; new
  first-class Developer variant + helpers + is_developer().
- All provider converters (anthropic / google / openai /
  openai-compatible — chat / responses / completion) handle Developer
  exhaustively and emit Warning::Unsupported on non-text parts via a
  local collapse_text_parts helper. google adds *_with_warnings entry
  point so prompt-conversion warnings bubble to GenerateResult.
- New services/inference/src/prompt_layout.{rs,test.rs}:
  PromptEnvelope, PromptSection(Kind/Source), CacheHint,
  PromptLayoutOptions, AnthropicSystemBlock(+CacheControl),
  PromptHashInputs, PROMPT_LAYOUT_NAMESPACE, plus
  put_/take_layout_options helpers storing fields as separate inner
  keys. coco-inference re-exports the new types.
- inference/client.rs::extract_system_text walks Vec<UserContentPart>
  via collapse_text_parts (consistent with the wire path).
- ai/{generate_text/prune_messages,prompt/convert}: Developer added
  to existing exhaustive matches.
- Migrate prior callers of `System { content: String }` (wrapping,
  token_estimation, cli output, compact tokens, auto-mode classifier
  prompt build) to the new Vec form and cover Developer.
- Layout adapter, routing tables, provider-side prompt_layout reads,
  and PromptStateInput hash-field expansion stay in the next slice.

— macOS test isolation fixes
- session: TranscriptMetadata.modified_at now millisecond-resolution
  with numeric sort; second-resolution timestamps tied on close-
  together appends and fell back to dir order.
- sandbox: canonicalize WritableRoot path before strip_prefix so
  /var/folders/... vs /private/var/folders/... matches the resolved
  gitdir on macOS; SandboxConfig grows an explicit Default impl that
  honors `serde(default = "default_true")` on allow_pty (derive
  Default produced false, contradicting docs and the seatbelt test).
- config-reload: track each file under both its original path and a
  canonical-parent variant so FSEvents-supplied /private/var/...
  event paths still match /var/... configured paths.

+++ coco-rs/core/tools/src/tools/bash.rs
@@ -1,3 +1,4 @@
+use coco_messages::ToolResult;
 use coco_shell::read_only::is_read_only_command;
 use coco_shell::sandbox::BypassRequest;
 use coco_shell::sandbox::SandboxConfig as ShellSandboxConfig;
@@ -16,7 +17,6 @@ use coco_tool_runtime::ValidationResult;
 use coco_types::ToolId;
 use coco_types::ToolInputSchema;
 use coco_types::ToolName;
-use coco_types::ToolResult;
 use serde_json::Value;
 use std::collections::HashMap;
 use std::path::PathBuf;
@@ -317,7 +317,7 @@ impl Tool for BashTool {
     /// the executor persists the full output to a tool-results file and
     /// only keeps a truncated snippet inline. We match TS exactly so
     /// cross-runtime sessions handle large outputs identically.
-    fn max_result_size_chars(&self) -> i32 {
+    fn max_result_size_chars(&self) -> i64 {
         30_000
     }
 
@@ -851,6 +851,19 @@ fn build_image_block(bytes: &[u8]) -> Value {
 /// or `(Some(path), bytes)` after persisting. Failure to write the
 /// file silently returns `(None, 0)` — persistence is best-effort
 /// and shouldn't break the tool result.
+///
+/// TODO(level-1-pipeline): this is a Bash-only stub of Level 1 of the
+/// Tool Result Budget plan. It diverges from TS in three ways:
+///   1. `temp_dir()` instead of session-scoped `<sessionDir>/tool-results/`
+///   2. Adds parallel `persistedOutputPath`/`persistedOutputSize` JSON
+///      fields instead of replacing `tool_result.content` with the
+///      `<persisted-output>` envelope (so the model still sees full stdout).
+///   3. No idempotency — every turn rewrites with a fresh nanosecond
+///      timestamp, breaking prompt-cache stability.
+///
+/// Replace with delegation to
+/// `coco_tool_runtime::tool_result_storage::maybe_persist_large_tool_result`
+/// once Phase 1 of `docs/coco-rs/tool-result-budget-plan.md` lands.
 pub(crate) fn maybe_persist_oversized_output(bytes: &[u8]) -> (Option<String>, usize) {
     // Match TS `maxResultSizeChars: 30_000` as the persistence trigger.
     // We threshold against raw bytes since TS uses char count and
